added ConsoleClient, fix get all credentials command arguments
